Fixes to work on ARM.

+++ lib/filesystem/scanner/walk.go
@@ -32,7 +32,7 @@ func makeRegularInode(stat *wsyscall.Stat_t) *filesystem.RegularInode {
 	inode.Mode = filesystem.FileMode(stat.Mode)
 	inode.Uid = stat.Uid
 	inode.Gid = stat.Gid
-	inode.MtimeSeconds = stat.Mtim.Sec
+	inode.MtimeSeconds = int64(stat.Mtim.Sec)
 	inode.MtimeNanoSeconds = int32(stat.Mtim.Nsec)
 	inode.Size = uint64(stat.Size)
 	return &inode
@@ -50,7 +50,7 @@ func makeSpecialInode(stat *wsyscall.Stat_t) *filesystem.SpecialInode {
 	inode.Mode = filesystem.FileMode(stat.Mode)
 	inode.Uid = stat.Uid
 	inode.Gid = stat.Gid
-	inode.MtimeSeconds = stat.Mtim.Sec
+	inode.MtimeSeconds = int64(stat.Mtim.Sec)
 	inode.MtimeNanoSeconds = int32(stat.Mtim.Nsec)
 	inode.Rdev = stat.Rdev
 	return &inode
